requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.3.1
dataclasses-json==0.6.3
distro==1.9.0
fastapi==0.122.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==2.4
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.0.8
langchain-core==1.0.4
langchain-ollama==1.0.0
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
ollama==0.6.1
openai==2.7.2
orjson==3.11.4
ormsgpack==1.12.0
packaging==23.2
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.1
redis==7.0.1
regex==2025.11.3
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlite-vec==0.1.6
starlette==0.50.0
tenacity==8.2.3
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.1.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.9.4
zstandard==0.25.0
